[plugins] Use separate module for plugins
