+++ plugins/__init__.py
@@ -1,0 +1,14 @@
+import asyncio
+
+from .database import SESSION
+
+try:
+    loop = asyncio.get_running_loop()
+except RuntimeError:  # 'RuntimeError: There is no current event loop...'
+    loop = None
+
+if loop and loop.is_running():
+    loop.create_task(SESSION.create_all())
+else:
+    print("Starting new event loop")
+    result = asyncio.run(SESSION.create_all())
