+++ plugins/fun.py
