+++ plugins/tickets/__init__.py
@@ -1,7 +1,5 @@
 from .cog import Tickets
-from .database import SESSION
 
 
 async def setup(bot):
-    await SESSION.create_all()
     await bot.add_cog(Tickets(bot))
