+++ plugins/tickets/models.py
@@ -3,7 +3,9 @@ from typing import Optional
 from sqlalchemy import select, update, delete
 from sqlalchemy.orm import selectinload
 
-from .database import SESSION, Panel, Form, Field
+from plugins.database import SESSION
+
+from .database import Panel, Form, Field
 
 
 class PanelModel:
